requirements.txt
asyncio
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
GPUtil
aiohttp
moviepy
natsort
opencv-python
Pillow
psutil
pytz
pyrofork==2.2.11
requests
tgcrypto
uvloop
yt_dlp
pymegatools
ffmpeg
